+++ derive_tests/tests/complex.rs
@@ -5,10 +5,7 @@ use validify::{Validate, ValidationError, ValidationErrors};
 
 fn validate_unique_username(username: &str) -> Result<(), ValidationError> {
     if username == "xXxShad0wxXx" {
-        return Err(ValidationError::new_field(
-            "first_name",
-            "terrible_username",
-        ));
+        return Err(ValidationError::new_field("terrible_username"));
     }
 
     Ok(())